+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 ## Unreleased
 
+### Note
+- 新しい設定項目"pgroonga.target"が追加されました.
+	- すでにnoteのtextのみのindexを貼っていてそのまま利用したい場合、設定は不要です.
+	- cwとtextのマルチカラムに対応する場合、**"pgroonga.target"を"cw_and_text"**に設定する必要があります.
+  - index	に関する詳細は #14730 を、configに関しては`.config/example.yml` または `.config/docker_example.yml`の'pgroonga'の節を参照願います
+
+
 ### General
--
 
 ### Client
 - Enhance: 投稿フォームの「迷惑になる可能性があります」のダイアログを表示する条件においてCWを考慮するように
@@ -9,7 +15,7 @@
 - Enhance: クライアントエラー画面の多言語対応
 
 ### Server
--
+- Enhance: pgroongaを用いた検索でcwと本文を合わせて, またAND検索、OR検索が可能になります
 
 
 ## 2025.2.0